Added support to read structArray from parquet (#547)
* Added parquet StructArray

* Added support for nested struct.

* Updated examples.

+++ src/io/parquet/write/binary/basic.rs
@@ -129,7 +129,7 @@ pub(crate) fn encode_delta<O: Offset>(
         if let Some(validity) = validity {
             let lengths = offsets
                 .windows(2)
-                .map(|w| (w[1] - w[0]).to_isize() as i32)
+                .map(|w| (w[1] - w[0]).to_isize() as i64)
                 .zip(validity.iter())
                 .flat_map(|(x, is_valid)| if is_valid { Some(x) } else { None });
             let length = offsets.len() - 1 - validity.null_count();
@@ -137,11 +137,11 @@ pub(crate) fn encode_delta<O: Offset>(
 
             delta_bitpacked::encode(lengths, buffer);
         } else {
-            let lengths = offsets.windows(2).map(|w| (w[1] - w[0]).to_isize() as i32);
+            let lengths = offsets.windows(2).map(|w| (w[1] - w[0]).to_isize() as i64);
             delta_bitpacked::encode(lengths, buffer);
         }
     } else {
-        let lengths = offsets.windows(2).map(|w| (w[1] - w[0]).to_isize() as i32);
+        let lengths = offsets.windows(2).map(|w| (w[1] - w[0]).to_isize() as i64);
         delta_bitpacked::encode(lengths, buffer);
     }
